requirements.txt
GitPython==3.1.45
spire-doc==13.8.0